+++ src/cli/CHANGELOG.md
@@ -13,6 +13,10 @@ Section ordering: Added, Changed, Fixed, Removed, Security, Deprecated.
 
 - Rename `Header` layout components (and summary `componentType` refs) to `Heading` when running `studioctl app upgrade v9`.
 
+### Fixed
+
+- Explain access errors during `studioctl app upgrade`
+
 ## [0.1.0-preview.19] - 2026-08-06
 
 ### Added